Compute asset hashes in post_upgrade
This removes the build-time hash computation and make it happen when
`for_each_asset` is called, effectively in `post_upgrade` (and `init`).

The interface is still a bit cumbersome (e.g. `for_each_asset` is still
callback based) but this makes sure that `main.rs` is not changed; the
focus here is the asset hashes. This is the first step in removing
`CANISTER_ID` from the build process; we want to inject it in the assets
directly, and this will change `main.rs` and `for_each_asset`.

I've tested this locally and assets (at least `index.js`) have the same
ic-certificate header as they have on `main`.
